+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Update: text in app
 - Update: views for 2 different type of calculations
 - Update: page for calculating nicotine now shows how many shots do you need to use
+- New: calculate strength view
 
 ## 2.0.3